order, beckend

+++ web/src/model/Customer.java
@@ -6,34 +6,43 @@ import java.util.Date;
 public class Customer extends User {
 	public ArrayList<Order> orders;
 	public ShoppingCart cart;
-	public int points;
+	public Double points;
 	public CustomerType type;
+
 	public ArrayList<Order> getOrders() {
 		return orders;
 	}
+
 	public void setOrders(ArrayList<Order> orders) {
 		this.orders = orders;
 	}
+
 	public ShoppingCart getCart() {
 		return cart;
 	}
+
 	public void setCart(ShoppingCart cart) {
 		this.cart = cart;
 	}
-	public int getPoints() {
+
+	public Double getPoints() {
 		return points;
 	}
-	public void setPoints(int points) {
+
+	public void setPoints(Double points) {
 		this.points = points;
 	}
+
 	public CustomerType getType() {
 		return type;
 	}
+
 	public void setType(CustomerType type) {
 		this.type = type;
 	}
+
 	public Customer(String username, String password, String name, String surname, Gender gender, Date dateOfBirth,
-			Role role, ArrayList<Order> orders, ShoppingCart cart, int points, CustomerType type) {
+			Role role, ArrayList<Order> orders, ShoppingCart cart, Double points, CustomerType type) {
 		super(username, password, name, surname, gender, dateOfBirth, role);
 		this.orders = orders;
 		this.cart = cart;
@@ -41,7 +50,7 @@ public class Customer extends User {
 		this.type = type;
 	}
 
-	public Customer(User user){
+	public Customer(User user) {
 		this.username = user.username;
 		this.password = user.password;
 		this.name = user.name;
@@ -49,13 +58,14 @@ public class Customer extends User {
 		this.gender = user.gender;
 		this.dateOfBirth = user.dateOfBirth;
 		this.role = user.role;
-		this.points = 0;
+		this.points = 0.0;
 		this.type = new CustomerType();
 		this.cart = new ShoppingCart();
 		this.orders = new ArrayList<Order>();
-		
+
 	}
-	
-	public Customer() {}
+
+	public Customer() {
+	}
 
 }
